fix(lint): eslint issues
