+++ src/components/clients/insurance-needs/advanced-income-components/advanced-income-types.ts
@@ -1,0 +1,7 @@
+export type DurationScenarioType =
+  | "custom"
+  | "childTurns18"
+  | "retirement"
+  | "lifetime";
+
+export type CalculationModeType = "income-multiplier" | "expense-based";
